refactor: move logout fun from user to auth
