+++ src/main/kotlin/learn_mate_it/dev/domain/auth/application/service/AuthService.kt
@@ -12,5 +12,6 @@ interface AuthService {
     fun signUp(request: SignUpRequest)
     fun signIn(request: SignInRequest): TokenResponse
     fun authenticateWithApple(request: AppleLoginRequest): Authentication
+    fun logout(refreshToken: String)
 
 }
